fixing bugs for posteffects

+++ GVRf/Framework/framework/src/main/jni/objects/textures/image.h
@@ -159,8 +159,8 @@ protected:
     std::mutex mUpdateLock;
     short   mType;
     short   mLevels;
-    short   mWidth;
-    short   mHeight;
+    int   mWidth;
+    int   mHeight;
     short   mDepth;
     short   mState;
     int     mImageSize;
